requirements.txt
Django==3.2.6
beautifulsoup4==4.8.2
requests==2.22.0
line-bot-sdk==1.20.0